CLIP ViTL14/requirement.txt
fastapi
uvicorn
pymongo
pillow
sentence-transformers
numpy
torch
protobuf
python-dotenv
python-multipart